validating android annotations
