+++ compiler/tests/org/jetbrains/jet/TestJdkKind.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 JetBrains s.r.o.
+ * Copyright 2010-2014 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,4 +19,5 @@ package org.jetbrains.jet;
 public enum TestJdkKind {
     MOCK_JDK,
     FULL_JDK,
+    ANDROID_API,
 }
